last changes

+++ modules/im/lib/V2/Controller/Chat/Message.php
@@ -7,6 +7,7 @@ use Bitrix\Im\V2\Chat;
 use Bitrix\Im\V2\Controller\BaseController;
 use Bitrix\Im\V2\Controller\Filter\CheckActionAccess;
 use Bitrix\Im\V2\Entity\View\ViewCollection;
+use Bitrix\Im\V2\Message\Delete\DeletionMode;
 use Bitrix\Im\V2\Message\Delete\DisappearService;
 use Bitrix\Im\V2\Message\Forward\ForwardService;
 use Bitrix\Im\V2\Message\MessageError;
@@ -263,10 +264,16 @@ class Message extends BaseController
 	/**
 	 * @restMethod im.v2.Chat.Message.delete
 	 */
-	public function deleteAction(\Bitrix\Im\V2\Message $message): ?bool
+	public function deleteAction(\Bitrix\Im\V2\MessageCollection $messages): ?bool
 	{
-		$service = new DeleteService($message);
-		$service->setMode(DeleteService::MODE_AUTO);
+		if ($messages->count() > MessageService::getMultipleActionMessageLimit())
+		{
+			$this->addError(new MessageError(MessageError::TOO_MANY_MESSAGES));
+
+			return null;
+		}
+
+		$service = new DeleteService($messages);
 		$result = $service->delete();
 
 		if (!$result->isSuccess())
@@ -284,9 +291,9 @@ class Message extends BaseController
 	 */
 	public function disappearAction(\Bitrix\Im\V2\Message $message, int $hours): ?bool
 	{
-		$deleteService = new DeleteService($message);
+		$deleteService = DeleteService::getInstanceByMessage($message);
 
-		if ($deleteService->canDelete() < DeleteService::DELETE_HARD)
+		if ($deleteService->canDelete($message->getId()) !== DeletionMode::Complete)
 		{
 			$this->addError(new MessageError(MessageError::ACCESS_DENIED));
